+++ CHANGELOG.md
@@ -1,2 +1,9 @@
-# 0.1.0
-* init release 
+# 0.2.0
+
+- Fixed `Some Errors.`
+- Music Slider.
+- Music Tracker.
+
+## 0.1.0
+
+- init release